refactor: modularize pixel editor effects
