+++ html/assets/js/pixel-editor/index.js
@@ -1,3 +1,9 @@
+import { applyAdjustments } from "./applyAdjustments.js";
+import { applyColorGlow } from "./applyColorGlow.js";
+import { applyBloom } from "./applyBloom.js";
+import { applyColorTuning } from "./applyColorTuning.js";
+import { applyHueRemap } from "./applyHueRemap.js";
+
 'use strict';
 
 export const LAYER_DEFAULTS = {
@@ -33,194 +39,8 @@ if (typeof window !== 'undefined') {
 }
 
 // Utility
-function clamp8(v){ return v<0?0:(v>255?255:v)|0; }
 function debounced(ms, fn){ let t; return (...a)=>{ clearTimeout(t); t=setTimeout(()=>fn(...a),ms); }; }
 function throttled(ms, fn){ let last=0, timer; return (...a)=>{ const now=Date.now(); const remain=ms-(now-last); if(remain<=0){ last=now; fn(...a); } else { clearTimeout(timer); timer=setTimeout(()=>{ last=Date.now(); fn(...a); }, remain); } }; }
-
-  // Adjustments
-  function applyAdjustments(ctx, w, h, bri, con, sat){
-    const img = ctx.getImageData(0,0,w,h);
-    const d = img.data;
-    const b = bri/100 * 255;
-    const c = (con/100)+1;
-    const intercept = 128*(1-c);
-    const s = sat/100;
-    for(let i=0;i<d.length;i+=4){
-      let r=d[i], g=d[i+1], bl=d[i+2];
-      r = c*r + intercept + b;
-      g = c*g + intercept + b;
-      bl = c*bl + intercept + b;
-      const avg=(r+g+bl)/3;
-      r = avg + (r-avg)*s;
-      g = avg + (g-avg)*s;
-      bl = avg + (bl-avg)*s;
-      d[i]=clamp8(r); d[i+1]=clamp8(g); d[i+2]=clamp8(bl);
-    }
-    ctx.putImageData(img,0,0);
-  }
-
-  function rgbToHsl(r,g,b){
-    r/=255; g/=255; b/=255;
-    const max=Math.max(r,g,b), min=Math.min(r,g,b);
-    let h=0,s=0; const l=(max+min)/2;
-    if(max!==min){
-      const d=max-min;
-      s = l>0.5 ? d/(2-max-min) : d/(max+min);
-      switch(max){
-        case r: h=(g-b)/d+(g<b?6:0); break;
-        case g: h=(b-r)/d+2; break;
-        case b: h=(r-g)/d+4; break;
-      }
-      h*=60;
-    }
-    return {h, s, l};
-  }
-
-  function hslToRgb(h,s,l){
-    const c=(1-Math.abs(2*l-1))*s, hp=h/60, x=c*(1-Math.abs((hp%2)-1));
-    let r1=0,g1=0,b1=0;
-    if(0<=hp&&hp<1){r1=c;g1=x;} else if(1<=hp&&hp<2){r1=x;g1=c;}
-    else if(2<=hp&&hp<3){g1=c;b1=x;} else if(3<=hp&&hp<4){g1=x;b1=c;}
-    else if(4<=hp&&hp<5){r1=x;b1=c;} else if(5<=hp&&hp<6){r1=c;b1=x;}
-    const m=l-c/2;
-    return {r:clamp8((r1+m)*255), g:clamp8((g1+m)*255), b:clamp8((b1+m)*255)};
-  }
-
-  const BUCKETS = [
-    {k:'R',h:0},
-    {k:'Y',h:60},
-    {k:'G',h:120},
-    {k:'C',h:180},
-    {k:'B',h:240},
-    {k:'M',h:300}
-  ];
-
-  function nearestBucket(hh){
-    let bk='R', bd=1e9;
-    for(const b of BUCKETS){
-      const diff=Math.min(Math.abs(hh-b.h), 360-Math.abs(hh-b.h));
-      if(diff<bd){ bd=diff; bk=b.k; }
-    }
-    return bk;
-  }
-
-  const BAND_HUE=[0,0,60,120,180,240,300];
-
-  function applyColorTuning(ctx, w, h, gains){
-    const img = ctx.getImageData(0,0,w,h);
-    const d=img.data;
-    for(let i=0;i<d.length;i+=4){
-      const r=d[i], g=d[i+1], bl=d[i+2];
-      const hsl=rgbToHsl(r,g,bl);
-      const band=nearestBucket(hsl.h);
-      const gain=(gains[band]||0)/100;
-      const sat= Math.max(0, Math.min(1, hsl.s*(1+gain)));
-      const rgb = hslToRgb(hsl.h, sat, hsl.l);
-      d[i]=rgb.r; d[i+1]=rgb.g; d[i+2]=rgb.b;
-    }
-    ctx.putImageData(img,0,0);
-  }
-
-  function applyHueRemap(ctx, w, h, mapping, globalStrength){
-    const img = ctx.getImageData(0,0,w,h);
-    const d=img.data;
-    for(let i=0;i<d.length;i+=4){
-      const r=d[i], g=d[i+1], bl=d[i+2];
-      const hsl=rgbToHsl(r,g,bl);
-      const srcBand = nearestBucket(hsl.h);
-      const cfg = mapping[srcBand];
-      if(!cfg) continue;
-      const targetSel = cfg.t;
-      const eff = Math.max(0, Math.min(1, globalStrength * cfg.s));
-      if(targetSel && targetSel>0 && eff>0){
-        const tgtHue = BAND_HUE[targetSel];
-        let dh = ((tgtHue - hsl.h + 540) % 360) - 180;
-        const newH = (hsl.h + dh*eff + 360) % 360;
-        const rgb = hslToRgb(newH, hsl.s, hsl.l);
-        d[i]=rgb.r; d[i+1]=rgb.g; d[i+2]=rgb.b;
-      }
-    }
-    ctx.putImageData(img,0,0);
-  }
-
-  // Shared offscreen canvas reused across operations
-  const scratchCanvas = typeof OffscreenCanvas !== 'undefined'
-    ? new OffscreenCanvas(1,1)
-    : (typeof document !== 'undefined' ? document.createElement('canvas') : null);
-  const scratchCtx = scratchCanvas ? scratchCanvas.getContext('2d') : null;
-
-  /**
-   * Apply hue-based glow to bright areas.
-   * @param {CanvasRenderingContext2D} ctx
-   * @param {number} w - canvas width
-   * @param {number} h - canvas height
- * @param {Object} glows - map of hue buckets to strength/range
- * @param {number} [threshold=0] - minimum lightness [0-1] required for glow
- * @param {number} [globalMult=1] - multiplier applied to all band strengths
- */
-  function applyColorGlow(ctx, w, h, glows, threshold=0, globalMult=1){
-    if(!glows || !Object.values(glows).some(v=>v.s>0) || !scratchCtx) return;
-    scratchCanvas.width = w;
-    scratchCanvas.height = h;
-    for(const [bandKey, cfg] of Object.entries(glows)){
-      const strength=(cfg.s||0)*globalMult;
-      if(strength<=0) continue;
-      const range=cfg.r||0;
-      scratchCtx.drawImage(ctx.canvas,0,0);
-      const img=scratchCtx.getImageData(0,0,w,h);
-      const d=img.data;
-      for(let i=0;i<d.length;i+=4){
-        const r=d[i], g=d[i+1], bl=d[i+2];
-        const hsl=rgbToHsl(r,g,bl);
-        const b=nearestBucket(hsl.h);
-        if(b!==bandKey || hsl.l<=threshold){ d[i]=d[i+1]=d[i+2]=0; continue; }
-        const t=(hsl.l - threshold)/(1 - threshold);
-        const eff=strength*t;
-        const newL=Math.min(1, hsl.l+eff);
-        const newS=Math.min(1, hsl.s+eff);
-        const rgb=hslToRgb(hsl.h,newS,newL);
-        d[i]=rgb.r; d[i+1]=rgb.g; d[i+2]=rgb.b;
-      }
-      scratchCtx.putImageData(img,0,0);
-      ctx.save();
-      if(range>0) ctx.filter=`blur(${range}px)`;
-      ctx.globalCompositeOperation='lighter';
-      ctx.globalAlpha=Math.min(1, strength);
-      ctx.drawImage(scratchCanvas,0,0);
-      ctx.restore();
-    }
-  }
-
-  /**
-   * Apply a bloom effect by blurring bright pixels.
-   * @param {CanvasRenderingContext2D} ctx
-   * @param {number} w - canvas width
-   * @param {number} h - canvas height
-   * @param {number} [threshold=200] - luminance cutoff [0-255] for bloom
-   * @param {number} [blurRadius=0] - blur radius in pixels
-   * @param {number} [alpha=0] - output alpha [0-1] for the bloom overlay
-   */
-  function applyBloom(ctx, w, h, threshold=200, blurRadius=0, alpha=0){
-    if(alpha<=0 || !scratchCtx) return;
-    scratchCanvas.width = w;
-    scratchCanvas.height = h;
-    scratchCtx.drawImage(ctx.canvas,0,0);
-    const img=scratchCtx.getImageData(0,0,w,h);
-    const d=img.data;
-    for(let i=0;i<d.length;i+=4){
-      const r=d[i], g=d[i+1], b=d[i+2];
-      const lum=0.2126*r + 0.7152*g + 0.0722*b;
-      if(lum<threshold){ d[i]=d[i+1]=d[i+2]=0; }
-    }
-    scratchCtx.putImageData(img,0,0);
-    ctx.save();
-    if(blurRadius>0) ctx.filter=`blur(${blurRadius}px)`;
-    ctx.globalCompositeOperation='lighter';
-    const easedAlpha=Math.pow(alpha,3); // cubic easing for finer control
-    ctx.globalAlpha=Math.min(1, easedAlpha);
-    ctx.drawImage(scratchCanvas,0,0);
-    ctx.restore();
-  }
 
   function kmeansRGB(rgba, count, k, seed=0){
     const pts=new Array(count);
